requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3
openai>=1.3.0
pinecone>=3.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
mistralai>=0.1.0
requests>=2.31.0
pydantic>=2.4.0
fastapi>=0.104.0
uvicorn>=0.24.0